+++ packages/paraglide/paraglide-next/src/app/index.client.tsx
@@ -4,3 +4,4 @@ export { createMiddleware } from "./middleware"
 export { createNavigation } from "./navigation"
 
 export { PrefixStrategy } from "./routing/prefixStrategy"
+export { initializeLanguage } from "./initializeLanguage.client.js"