+++ packages/paraglide/paraglide-next/src/app/index.server.tsx
@@ -5,3 +5,5 @@ export { createMiddleware } from "./middleware"
 export { createNavigation } from "./navigation"
 
 export { PrefixStrategy } from "./routing/prefixStrategy"
+
+export { initializeLanguage } from "./initializeLanguage.server.js"