+++ development-projects/inlang-nextjs/CHANGELOG.md
@@ -1,5 +1,25 @@
 # next-js-testapp
 
+## 0.1.5
+
+### Patch Changes
+
+- @inlang/cli@2.15.1
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [5b8c053]
+  - @inlang/cli@2.15.0
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [cd9a3e1]
+  - @inlang/cli@2.14.2
+
 ## 0.1.2
 
 ### Patch Changes