+++ packages/paraglide/paraglide-sveltekit/example/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @inlang/paraglide-sveltekit-example
 
+## 0.1.29
+
+### Patch Changes
+
+- @inlang/paraglide-sveltekit@0.6.10
+
+## 0.1.28
+
+### Patch Changes
+
+- @inlang/paraglide-sveltekit@0.6.9
+
+## 0.1.27
+
+### Patch Changes
+
+- @inlang/paraglide-sveltekit@0.6.8
+
 ## 0.1.26
 
 ### Patch Changes